move ToggleButton to a subcomponent
